refactor: restructure project
+ move state to a parent app component
+ move view components to /components
+ unify data fetching, pre-rendering and rendering in app component
