+++ src/commands/diff/View.js
@@ -1,0 +1,3 @@
+import DiffContainer from "components/DiffContainer"
+
+export default DiffContainer
