+++ src/components/Preview.js
@@ -2,7 +2,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import {Box, Text} from "ink"
 
-import {calculatePreviewWindow} from "./utils"
+import {calculatePreviewWindow} from "commands/diff"
 
 export default function Preview({preview, width, height, previewPosition}) {
   if (preview.length === 0 || width === 0) {
